% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freebase_functions.R
\docType{package}
\name{freebase_googleAuthR}
\alias{freebase_googleAuthR}
\alias{freebase_googleAuthR-package}
\title{Freebase Search
Find Freebase entities using textual queries and other constraints.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:13:01
filename: /Users/mark/dev/R/autoGoogleAPI/googlefreebasev1.auto/R/freebase_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}

