% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freebase_objects.R
\name{ReconcileCandidate.notable}
\alias{ReconcileCandidate.notable}
\title{ReconcileCandidate.notable Object}
\usage{
ReconcileCandidate.notable(id = NULL, name = NULL)
}
\arguments{
\item{id}{MID of notable category}

\item{name}{Name of notable category in specified language}
}
\value{
ReconcileCandidate.notable object
}
\description{
ReconcileCandidate.notable Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Type or profession the candidate is notable for.
}
\seealso{
Other ReconcileCandidate functions: \code{\link{ReconcileCandidate}}
}

