% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freebase_objects.R
\name{ReconcileCandidate}
\alias{ReconcileCandidate}
\title{Freebase Search Objects 
Find Freebase entities using textual queries and other constraints.}
\usage{
ReconcileCandidate(ReconcileCandidate.notable = NULL, confidence = NULL,
  lang = NULL, mid = NULL, name = NULL, notable = NULL)
}
\arguments{
\item{ReconcileCandidate.notable}{The \link{ReconcileCandidate.notable} object or list of objects}

\item{confidence}{Percentage likelihood that this candidate is the unique matching entity}

\item{lang}{Language code that candidate and notable names are displayed in}

\item{mid}{Freebase MID of candidate entity}

\item{name}{Freebase name of matching entity in specified language}

\item{notable}{Type or profession the candidate is notable for}
}
\value{
ReconcileCandidate object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:13:01
filename: /Users/mark/dev/R/autoGoogleAPI/googlefreebasev1.auto/R/freebase_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ReconcileCandidate Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ReconcileCandidate functions: \code{\link{ReconcileCandidate.notable}}
}

