% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freebase_objects.R
\name{ReconcileGet.costs}
\alias{ReconcileGet.costs}
\title{ReconcileGet.costs Object}
\usage{
ReconcileGet.costs(hits = NULL, ms = NULL)
}
\arguments{
\item{hits}{Total number of hits found}

\item{ms}{Total milliseconds spent}
}
\value{
ReconcileGet.costs object
}
\description{
ReconcileGet.costs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Server costs for reconciling.
}
\seealso{
Other ReconcileGet functions: \code{\link{ReconcileGet.warning}},
  \code{\link{ReconcileGet}}
}

