% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freebase_objects.R
\name{ReconcileGet}
\alias{ReconcileGet}
\title{ReconcileGet Object}
\usage{
ReconcileGet(ReconcileGet.costs = NULL, ReconcileGet.warning = NULL,
  candidate = NULL, costs = NULL, match = NULL, warning = NULL)
}
\arguments{
\item{ReconcileGet.costs}{The \link{ReconcileGet.costs} object or list of objects}

\item{ReconcileGet.warning}{The \link{ReconcileGet.warning} object or list of objects}

\item{candidate}{If filled, then the listed candidates are potential matches, and such should be evaluated by a more discerning algorithm or human}

\item{costs}{Server costs for reconciling}

\item{match}{If filled, this entity is guaranteed to match at requested confidence probability (default 99%)}

\item{warning}{If filled, then there were recoverable problems that affected the request}
}
\value{
ReconcileGet object
}
\description{
ReconcileGet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ReconcileGet functions: \code{\link{ReconcileGet.costs}},
  \code{\link{ReconcileGet.warning}}
}

